+++ satella/coding/structures/dictionaries/dict_object.py
@@ -102,6 +102,12 @@ class DictObject(tp.MutableMapping[str, T]):
         else:
             return True
 
+    def get(self, k: str, v: tp.Optional[tp.Any] = None) -> tp.Any:
+        try:
+            return self.__data[k]
+        except KeyError:
+            return v
+
 
 def apply_dict_object(v: tp.Union[tp.Any, tp.Dict[str, T]]) -> tp.Union[DictObject, tp.Any]:
     """